+++ measure.h
@@ -17,7 +17,7 @@ extern int T_meas; // milli deg C
 
 //Read ADC10MEM register into this buffer
 extern float read_buff[CAPACITY]; //to store past values
-extern float temp_read_buff[N_POINTS]; //to compute avg value on every loop
+extern int temp_read_buff[N_POINTS]; //to compute avg value on every loop
 
 // Store index of array
 extern int index;
